fix AnimationSkeletonController transform attribute name collision risk
`AnimationSkeletonController` requires transform attributes to store such
values for computations, but is among others a template of the transform type.

If two controllers (for different animation types) that share the same
attribute names had different transform types, the dynamic cast in
`AttributeContainerGen::get_attribute` would fail and return a null pointer.
We could check whether or not that pointer is null and e.g. warn if it is,
but this then requires the user to figure out what the issue is.

By default, the names are ensured to be unique between transform types
by integrating a suffix, which is the prettified name of that type.

The caller is free to provide other names, and given that the parameter names
contain "unique", it should be pretty clear that
it doesn't accept identical names between types.
Delegating the responsibility of attribute management to the caller
is also coherent to other paradigms within the program.